Move reading plan file migration to upgradeSharedPreferences() in BibleApplication
